Update ArrayMulp.cpp
When NM size matrix A and MK size matrix B are given, complete a program that multiplies two matrices.

Input The first line is given the sizes N and M of the matrix A. From the second line to the N line, M elements of matrix A are given in order. The next line is given the size M and K of the matrix B. Subsequently, K elements of matrix B are given to M lines in order. N, M, and K are less than or equal to 100, and the elements of the matrix are integers whose absolute value is less than or equal to 100.