Create simple dining-house program
This program is implemented in Python, and the contents of the buttons or other contents are written in Korean, not English. The contents written in Korean are simple contents, so if you need to refer to the code, check the contents through the translator. And this program was written using Jupiter Notebook.