Create ArraySquare
A matrix A of size N*N is given. At this time, write a program that seeks A's B-square. Since the number can be very large, print out the remainder of each element of ABB divided by 1,000.

Input The first line is given the sizes N and B of the matrix. (2 ≤ N ≤ 5, 1 ≤ B ≤ 100,000,000,000)

From the second line, N lines are given each element of the matrix. Each element of the matrix is a natural number less than or equal to 1,000 or equal to zero.

Output Outputs the results of the B-square matrix A over N lines from the first line.