Create LargeNum_Square
I want to know the number of square B of natural number A. However, since the number you want to obtain can be very large, write a program that will divide it by C and find the rest.

Input In the first line, A, B, and C are given in order with spaces between them. A, B and C are all natural numbers of 2,147,483,647.